pyproject.toml
[project]
name = "aura"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "alembic>=1.18.1",
    "fastapi>=0.128.0",
    "structlog>=25.0.0",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "protobuf>=6.33.5",
    "grpcio-health-checking>=1.76.0",
    "langchain-mistralai>=1.1.1",
    "mypy-protobuf>=5.0.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "pynacl>=1.5.0",
    # OpenTelemetry dependencies
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-instrumentation-grpc>=0.45b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.45b0",
    "opentelemetry-instrumentation-langchain>=0.1.0",
    "httpx>=0.27.0",
    "litellm>=1.63.0",
    "jinja2>=3.1.4",
    "dspy-ai>=2.0.0",
    "prometheus-client>=0.21.1",
    # Solana dependencies for crypto payments
    "solana>=0.34.0",
    "solders>=0.21.0",
    # Encryption for secret storage
    "cryptography>=43.0.0",
    "nats-py>=2.9.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "bandit>=1.8.0",
    "pre-commit>=4.0.0",
    "types-requests",
    "types-setuptools",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "aiogram>=3.24.0",
]

[tool.ruff]
target-version = "py313"
line-length = 88
exclude = [
    "**/proto/**",  # Auto-generated protobuf files
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["config", "db", "embeddings", "llm_strategy", "proto"]

[tool.mypy]
python_version = "3.12"
explicit_package_bases = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_return_any = false
warn_unused_configs = true
plugins = ["pydantic.mypy"]
exclude = ["proto/.*", "tests/.*"]

[[tool.mypy.overrides]]
module = [
    "solana.*",
    "solders.*",
    "grpc.*",
    "google.protobuf.*",
    "dspy.*",
    "prometheus_client.*",
    "nats.*",
    "aiogram.*",
    "sqlalchemy.*",
    "alembic.*",
    "pgvector.*",
    "grpc_health.*",
    "opentelemetry.*",
    "litellm.*",
]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101"] # B101 is assert_used, often okay in some contexts but bandit defaults can be strict.

[[tool.mypy.overrides]]
module = ["proto.*", "aura.*", "src.proto.*"]
ignore_errors = true

# [[tool.uv.index]]
# url = "https://buf.build/gen/python"
